+++ src/main/java/forestry/core/render/RenderCrateItem.java
@@ -22,10 +22,7 @@ import net.minecraft.entity.item.EntityItem;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemBlock;
 import net.minecraft.item.ItemStack;
-import net.minecraft.util.IIcon;
 import net.minecraft.util.ResourceLocation;
-
-import net.minecraftforge.client.IItemRenderer;
 
 import org.lwjgl.opengl.GL11;
 
@@ -105,9 +102,9 @@ public class RenderCrateItem implements IItemRenderer {
 
 			if (renderItem.renderWithColor) {
 				int color = stack.getItem().getColorFromItemStack(stack, pass);
-				float c1 = (float) (color >> 16 & 255) / 255.0F;
-				float c2 = (float) (color >> 8 & 255) / 255.0F;
-				float c3 = (float) (color & 255) / 255.0F;
+				float c1 = (color >> 16 & 255) / 255.0F;
+				float c2 = (color >> 8 & 255) / 255.0F;
+				float c3 = (color & 255) / 255.0F;
 
 				GL11.glColor4f(c1, c2, c3, 1.0F);
 			}
@@ -141,7 +138,7 @@ public class RenderCrateItem implements IItemRenderer {
 		float offsetZ = PIXEL + 0.021875F;
 
 		GL11.glRotatef((((float) entity.age + 1.0F) / 20.0F + entity.hoverStart) * (180F / (float) Math.PI), 0.0F, 1.0F, 0.0F);
-		GL11.glTranslatef(-0.5F, -0.25F, -(offsetZ * (float) iterations / 2.0F));
+		GL11.glTranslatef(-0.5F, -0.25F, -(offsetZ * iterations / 2.0F));
 
 		for (int count = 0; count < iterations; ++count) {
 			if (count > 0) {
@@ -183,9 +180,9 @@ public class RenderCrateItem implements IItemRenderer {
 
 			if (renderItem.renderWithColor) {
 				int color = stack.getItem().getColorFromItemStack(stack, pass);
-				float c1 = (float) (color >> 16 & 255) / 255.0F;
-				float c2 = (float) (color >> 8 & 255) / 255.0F;
-				float c3 = (float) (color & 255) / 255.0F;
+				float c1 = (color >> 16 & 255) / 255.0F;
+				float c2 = (color >> 8 & 255) / 255.0F;
+				float c3 = (color & 255) / 255.0F;
 
 				GL11.glColor4f(c1, c2, c3, 1.0F);
 			}
@@ -214,7 +211,7 @@ public class RenderCrateItem implements IItemRenderer {
 
 					float f13 = 0.76F;
 					float f14 = 0.125F;
-					float f15 = (float) (Minecraft.getSystemTime() % 3000L) / 3000.0F * 8.0F;
+					float f15 = Minecraft.getSystemTime() % 3000L / 3000.0F * 8.0F;
 					GL11.glColor4f(0.5F * f13, 0.25F * f13, 0.8F * f13, 1.0F);
 
 					GL11.glMatrixMode(GL11.GL_TEXTURE);
@@ -231,7 +228,7 @@ public class RenderCrateItem implements IItemRenderer {
 					GL11.glPushMatrix();
 					{
 						GL11.glScalef(f14, f14, f14);
-						f15 = (float) (Minecraft.getSystemTime() % 4873L) / 4873.0F * 8.0F;
+						f15 = Minecraft.getSystemTime() % 4873L / 4873.0F * 8.0F;
 						GL11.glTranslatef(-f15, 0.0F, 0.0F);
 						GL11.glRotatef(10.0F, 0.0F, 0.0F, 1.0F);
 						ItemRenderer.renderItemIn2D(tessellator, 0.0F, 0.0F, 1.0F, 1.0F, 255, 255, PIXEL);
@@ -319,9 +316,9 @@ public class RenderCrateItem implements IItemRenderer {
 
 					if (renderItem.renderWithColor) {
 						int color = contained.getItem().getColorFromItemStack(contained, pass);
-						float c1 = (float) (color >> 16 & 255) / 255.0F;
-						float c2 = (float) (color >> 8 & 255) / 255.0F;
-						float c3 = (float) (color & 255) / 255.0F;
+						float c1 = (color >> 16 & 255) / 255.0F;
+						float c2 = (color >> 8 & 255) / 255.0F;
+						float c3 = (color & 255) / 255.0F;
 
 						GL11.glColor4f(c1, c2, c3, 1.0F);
 					}
